Add initial interfaces to complete the test, but it still breaks

+++ src/test/java/hu/gvasko/codekata/datamunging/DataMungingUtil.java
@@ -1,5 +1,9 @@
 package hu.gvasko.codekata.datamunging;
 
+import hu.gvasko.stringtable.StringRecord;
+
+import java.util.List;
+
 /**
  * Utility functions to delay design decisions
  * Created by gvasko on 2015.05.06..
@@ -10,7 +14,7 @@ public class DataMungingUtil {
         return records.stream().min( (rec1, rec2) -> Integer.compare(getDiff(rec1, f1, f2), getDiff(rec2, f1, f2)) ).get();
     }
 
-    private static float getDiff(StringRecord rec, String f1, String f2) {
+    private static int getDiff(StringRecord rec, String f1, String f2) {
         return Math.abs(Integer.parseInt(rec.get(f1)) - Integer.parseInt(rec.get(f2)));
     }
